refactor(paxos): disentangle KV replicas from Paxos
Makes it possible to substitute Paxos with other sequencing protocols.
