+++ hydro_test/examples/paxos.rs
@@ -3,6 +3,7 @@ use std::sync::Arc;
 use hydro_deploy::gcp::GcpNetwork;
 use hydro_deploy::{Deployment, Host};
 use hydro_lang::deploy::TrybuildHost;
+use hydro_test::cluster::paxos::PaxosConfig;
 use tokio::sync::RwLock;
 
 type HostCreator = Box<dyn Fn(&mut Deployment) -> Arc<dyn Host>>;
@@ -43,13 +44,15 @@ async fn main() {
 
     let (proposers, acceptors, clients, replicas) = hydro_test::cluster::paxos_bench::paxos_bench(
         &builder,
-        f,
         num_clients_per_node,
         median_latency_window_size,
         checkpoint_frequency,
-        i_am_leader_send_timeout,
-        i_am_leader_check_timeout,
-        i_am_leader_check_timeout_delay_multiplier,
+        PaxosConfig {
+            f,
+            i_am_leader_send_timeout,
+            i_am_leader_check_timeout,
+            i_am_leader_check_timeout_delay_multiplier,
+        },
     );
 
     let rustflags = "-C opt-level=3 -C codegen-units=1 -C strip=none -C debuginfo=2 -C lto=off";
