+++ hydro_test/src/cluster/mod.rs
@@ -1,8 +1,10 @@
+pub mod bench_client;
 pub mod compute_pi;
+pub mod kv_replica;
 pub mod many_to_many;
 pub mod map_reduce;
 pub mod paxos;
 pub mod paxos_bench;
-pub mod paxos_kv;
+pub mod paxos_with_client;
 pub mod simple_cluster;
 pub mod two_pc;
